// src/core/domain/enums/error-codes.ts
export enum ErroCodes {
    USER_UNAUTHORIZED = "USER_UNAUTHORIZED",
    USER_PASSWORD_REQUIRED = "USER_PASSWORD_REQUIRED",
    USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
    TOKEN_NOT_PROVIDED = "TOKEN_NOT_PROVIDED",
    TOKEN_INVALID = "TOKEN_INVALID",
    TOKEN_EXPIRED = "TOKEN_EXPIRED",
    TOKEN_NOT_FOUND = "TOKEN_NOT_FOUND",
    USER_NAME_REQUIRED = "USER_NAME_REQUIRED",
    USER_EMAIL_REQUIRED = "USER_EMAIL_REQUIRED",
    USER_NAME_MIN_LENGTH = "USER_NAME_MIN_LENGTH",
    USER_EMAIL_MIN_LENGTH = "USER_EMAIL_MIN_LENGTH",
    USER_PASSWORD_MIN_LENGTH = "USER_PASSWORD_MIN_LENGTH",
    USER_EMAIL_INVALID = "USER_EMAIL_INVALID",
    USER_NAME_INVALID = "USER_NAME_INVALID",
    USER_NOT_FOUND = "USER_NOT_FOUND",
    USER_ID_REQUIRED = "USER_ID_REQUIRED",
    USER_EMAIL_NOT_FOUND = "USER_EMAIL_NOT_FOUND",
    CLIENT_NAME_REQUIRED = "CLIENT_NAME_REQUIRED",
    CLIENT_EMAIL_REQUIRED = "CLIENT_EMAIL_REQUIRED",
    CLIENT_EMAIL_INVALID = "CLIENT_EMAIL_INVALID",
    CLIENT_DOCUMENT_REQUIRED = "CLIENT_DOCUMENT_REQUIRED",
    CLIENT_ALREADY_EXISTS = "CLIENT_ALREADY_EXISTS",
    CLIENT_NOT_FOUND = "CLIENT_NOT_FOUND",
    CLIENT_ID_REQUIRED = "CLIENT_ID_REQUIRED",
    CLIENT_PHONE_REQUIRED = "CLIENT_PHONE_REQUIRED",
    CLIENT_PAGE_REQUIRED = "CLIENT_PAGE_REQUIRED",
    CLIENT_PAGE_MAX = "CLIENT_PAGE_MAX",
    CLIENT_LIMIT_REQUIRED = "CLIENT_LIMIT_REQUIRED",
    CLIENT_LIMIT_MAX = "CLIENT_LIMIT_MAX",
    CLIENT_PAGE_INVALID = "CLIENT_PAGE_INVALID",
    CLIENT_LIMIT_INVALID = "CLIENT_LIMIT_INVALID",
    VENDOR_NAME_REQUIRED = "VENDOR_NAME_REQUIRED",
    VENDOR_EMAIL_REQUIRED = "VENDOR_EMAIL_REQUIRED",
    VENDOR_EMAIL_INVALID = "VENDOR_EMAIL_INVALID",
    VENDOR_DOCUMENT_REQUIRED = "VENDOR_DOCUMENT_REQUIRED",
    VENDOR_ALREADY_EXISTS = "VENDOR_ALREADY_EXISTS",
    VENDOR_NOT_FOUND = "VENDOR_NOT_FOUND",
    VENDOR_ID_REQUIRED = "VENDOR_ID_REQUIRED",
    VENDOR_PHONE_REQUIRED = "VENDOR_PHONE_REQUIRED",
    VENDOR_PAGE_REQUIRED = "VENDOR_PAGE_REQUIRED",
    VENDOR_PAGE_MAX = "VENDOR_PAGE_MAX",
    VENDOR_LIMIT_REQUIRED = "VENDOR_LIMIT_REQUIRED",
    VENDOR_LIMIT_MAX = "VENDOR_LIMIT_MAX",
    VENDOR_PAGE_INVALID = "VENDOR_PAGE_INVALID",
    VENDOR_LIMIT_INVALID = "VENDOR_LIMIT_INVALID",
    PRODUCT_NAME_REQUIRED = "PRODUCT_NAME_REQUIRED",
    PRODUCT_COST_PRICE_REQUIRED = "PRODUCT_COST_PRICE_REQUIRED",
    PRODUCT_SELL_PRICE_REQUIRED = "PRODUCT_SELL_PRICE_REQUIRED",
    PRODUCT_ALREADY_EXISTS = "PRODUCT_ALREADY_EXISTS",
    PRODUCT_NOT_FOUND = "PRODUCT_NOT_FOUND",
    PRODUCT_ID_REQUIRED = "PRODUCT_ID_REQUIRED",
    PRODUCT_PAGE_REQUIRED = "PRODUCT_PAGE_REQUIRED",
    PRODUCT_PAGE_MAX = "PRODUCT_PAGE_MAX",
    PRODUCT_LIMIT_REQUIRED = "PRODUCT_LIMIT_REQUIRED",
    PRODUCT_LIMIT_MAX = "PRODUCT_LIMIT_MAX",
    PRODUCT_PAGE_INVALID = "PRODUCT_PAGE_INVALID",
    PRODUCT_LIMIT_INVALID = "PRODUCT_LIMIT_INVALID",
    SERVICE_NAME_REQUIRED = "SERVICE_NAME_REQUIRED",
    SERVICE_PRICE_REQUIRED = "SERVICE_PRICE_REQUIRED",
    SERVICE_ALREADY_EXISTS = "SERVICE_ALREADY_EXISTS",
    SERVICE_NOT_FOUND = "SERVICE_NOT_FOUND",
    SERVICE_ID_REQUIRED = "SERVICE_ID_REQUIRED",
    SERVICE_PAGE_REQUIRED = "SERVICE_PAGE_REQUIRED",
    SERVICE_PAGE_MAX = "SERVICE_PAGE_MAX",
    SERVICE_LIMIT_REQUIRED = "SERVICE_LIMIT_REQUIRED",
    SERVICE_LIMIT_MAX = "SERVICE_LIMIT_MAX",
    SERVICE_PAGE_INVALID = "SERVICE_PAGE_INVALID",
    SERVICE_LIMIT_INVALID = "SERVICE_LIMIT_INVALID",
    SERVICE_ORDER_CLIENT_ID_REQUIRED = "SERVICE_ORDER_CLIENT_ID_REQUIRED",
    SERVICE_ORDER_DESCRIPTION_REQUIRED = "SERVICE_ORDER_DESCRIPTION_REQUIRED",
    SERVICE_ORDER_SERVICE_ID_REQUIRED = "SERVICE_ORDER_SERVICE_ID_REQUIRED",
    SERVICE_ORDER_PRODUCT_ID_REQUIRED = "SERVICE_ORDER_PRODUCT_ID_REQUIRED",
    SERVICE_ORDER_STATUS_REQUIRED = "SERVICE_ORDER_STATUS_REQUIRED",
    SERVICE_ORDER_INVALID_STATUS = "SERVICE_ORDER_INVALID_STATUS",
    SERVICE_ORDER_NOT_FOUND = "SERVICE_ORDER_NOT_FOUND",
    SERVICE_ORDER_ID_REQUIRED = "SERVICE_ORDER_ID_REQUIRED",
    SERVICE_ORDER_PAGE_REQUIRED = "SERVICE_ORDER_PAGE_REQUIRED",
    SERVICE_ORDER_PAGE_MAX = "SERVICE_ORDER_PAGE_MAX",
    SERVICE_ORDER_LIMIT_REQUIRED = "SERVICE_ORDER_LIMIT_REQUIRED",
    SERVICE_ORDER_LIMIT_MAX = "SERVICE_ORDER_LIMIT_MAX",
    SERVICE_ORDER_PAGE_INVALID = "SERVICE_ORDER_PAGE_INVALID",
    SERVICE_ORDER_LIMIT_INVALID = "SERVICE_ORDER_LIMIT_INVALID",
}